Moved invalid QR code dialog to the qr code scan screen.
